requirements.txt
pyyaml~=6.0.1
transformers~=4.31
torch~=2.0.1
accelerate~=0.21.0
xformers